Create Action menu by JS
