+++ src/index.html
@@ -45,33 +45,14 @@
       </div>
     </section>
     <!--* Action Menu -->
-    <div class="action-menu">
-      <div class="view-button">
-        <div class="button-text">View</div>
-        <div class="icon-wrapper">
-          <img src="./assets/icons/view-icon.svg" alt="View icon" />
-        </div>
-      </div>
-      <div class="edit-button">
-        <div class="button-text">Edit</div>
-        <div class="icon-wrapper">
-          <img src="./assets/icons/edit-icon.svg" alt="Edit icon" />
-        </div>
-      </div>
-      <div class="delete-button">
-        <div class="button-text delete">Delete</div>
-        <div class="icon-wrapper">
-          <img src="./assets/icons/delete-icon.svg" alt="Delete icon" />
-        </div>
-      </div>
-    </div>
+    <div class="action-menu"></div>
     <!--* Modal Overlay -->
     <div class="modal-overlay"></div>
     <!--* Customer Modal will be created in JS -->
     <div class="modal customer-modal"></div>
     <!--* View Customer Modal will be created in JS -->
     <div class="modal view-customer-modal"></div>
-    <!-- Delete confirmation modal will be created by JS-->
+    <!--* Delete confirmation modal will be created by JS-->
     <div class="modal delete-confirmation-modal"></div>
     <script src="./js/main.js" type="module"></script>
   </body>
