+++ src/js/models/customer.js
@@ -1,0 +1,25 @@
+class Customer {
+  constructor(id, name, status, rate, balance, deposit, description) {
+    this.id = id;
+    this.name = name;
+    this.status = status;
+    this.rate = rate;
+    this.balance = balance;
+    this.deposit = deposit;
+    this.description = description;
+  }
+
+  toJSON() {
+    return JSON.stringify({
+      id: this.id,
+      name: this.name,
+      status: this.status,
+      rate: this.rate,
+      balance: this.balance,
+      deposit: this.deposit,
+      description: this.description,
+    });
+  }
+}
+
+export default Customer;
