+++ src/js/utils/helpers.js
@@ -77,6 +77,37 @@ function sortCustomersByName(customers, currentSortingState) {
   }
 }
 
+//* Open and close modal functions
+function openModal(modal) {
+  const modalElement = modal;
+  const modalOverlay = document.querySelector('.modal-overlay');
+  modalElement.classList.add('open'); // Show the modal
+  modalOverlay.classList.add('open'); // Show the overlay
+}
+
+function closeModal(modal) {
+  const modalElement = modal;
+  const modalOverlay = document.querySelector('.modal-overlay');
+  modalElement.classList.remove('open'); // Hide the modal
+  while (modalElement.firstChild) {
+    modalElement.removeChild(modalElement.firstChild);
+  }
+
+  modalOverlay.classList.remove('open'); // Hide the overlay
+}
+
+// Add/Edit form validation
+function checkFormValidity() {
+  const confirmButton = document.querySelector('.confirm-button');
+  const nameInput = document.getElementById('name-input');
+  const rateInput = document.getElementById('rate-input');
+  const balanceInput = document.getElementById('balance-input');
+  const depositInput = document.getElementById('deposit-input');
+  const isFormValid =
+    nameInput.value && rateInput.value && balanceInput.value && depositInput.value;
+  confirmButton.disabled = !isFormValid;
+}
+
 export {
   hasNumbers,
   enforceMaxLength,
@@ -86,4 +117,7 @@ export {
   combineAndRemoveDuplicates,
   getActionMenuPosition,
   sortCustomersByName,
+  openModal,
+  closeModal,
+  checkFormValidity,
 };
