+++ src/js/utils/state.js
@@ -1,0 +1,6 @@
+const state = {
+  isAddMode: false,
+  currentCustomer: null,
+};
+
+export default state;
